backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
mysqlclient==2.2.0
djangorestframework==3.14.0
pytz==2023.3
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2